extracts Note component from MainScreen
